production/requirements.txt
numpy==1.23.5
Pillow==9.4.0
scikit_learn==1.2.1
tensorflow==2.13.0
tensorflow_hub==0.13.0
tensorflow_intel==2.13.0